Consistent hash algorithm modified int32_t to uint32_t

+++ servant/servant/EndpointManager.h
@@ -387,28 +387,28 @@ private:
     /*
      * 根据hash值选取一个结点
      */
-    AdapterProxy* getHashProxy(int64_t hashCode,  bool bConsistentHash = false);
+    AdapterProxy* getHashProxy(uint32_t hashCode,  bool bConsistentHash = false);
 
 
     /*
      * 根据hash值按取模方式，从正常节点中选取一个结点
      */
-    AdapterProxy* getHashProxyForNormal(int64_t hashCode);
+    AdapterProxy* getHashProxyForNormal(uint32_t hashCode);
 
     /*
      * 根据hash值按一致性hash方式，从正常节点中选取一个结点
      */
-    AdapterProxy* getConHashProxyForNormal(int64_t hashCode);
+    AdapterProxy* getConHashProxyForNormal(uint32_t hashCode);
 
     /*
      * 根据hash值按取模方式，从静态权重节点中选取一个结点
      */
-    AdapterProxy* getHashProxyForWeight(int64_t hashCode, bool bStatic, vector<size_t> &vRouterCache);
+    AdapterProxy* getHashProxyForWeight(uint32_t hashCode, bool bStatic, vector<size_t> &vRouterCache);
 
     /*
      * 根据hash值按一致性hash方式，从静态权重节点中选取一个结点
      */
-    AdapterProxy* getConHashProxyForWeight(int64_t hashCode, bool bStatic);
+    AdapterProxy* getConHashProxyForWeight(uint32_t hashCode, bool bStatic);
 
     /*
      * 判断静态权重节点是否有变化
